Started over
I have started the project over after becoming
unhappy with how the form looked. I've decided to
style the form similarly to the speaker form shown
during some of the exercises but I have and will
be heavily tweaking it to my own style and perhaps
adding features.

I've deleted the original files and replaced them
with the new set under the same names. I have also
replaced the images. I will keep the fonts file as
that is something I will also be playing around
with.

I am liking it much better already. It is also
(due to the exercise) a mobile first approach so
the form will be built to be very amenable to both
mobile and PC browsers.